backend/requirements.txt
fastapi
uvicorn
python-multipart
moviepy
openai-whisper
easyocr
sentence-transformers
torch
pillow
numpy
